chore: move plugins and rules into lib folder, add package exports map
Moves all published config entry points into lib/, updates package.json
main/files/exports to match, and picks up pending pnpm workspace/dependency
changes (pnpm-workspace.yaml, exact dep pins, packageManager bump).
