+++ .npmignore
@@ -1,8 +1,5 @@
 *
-!*.js
-commitlint.config.js
-eslint.config.js
-!rules
+!lib
 !README.md
 !package.json
 !pnpm-lock.yaml
