--- .npmrc
@@ -1,2 +1,0 @@
-public-hoist-pattern[]=@commitlint*
-public-hoist-pattern[]=commitlint
