+++ .prettierignore
@@ -1,3 +1,1 @@
-.cache
 node_modules
-.pnpm-store
