+++ lib/a11y.js
@@ -7,7 +7,7 @@ module.exports = [
     languageOptions: jsxA11y.flatConfigs.recommended.languageOptions,
     rules: {
       ...jsxA11y.flatConfigs.recommended.rules,
-      ...require('./lib/rules/a11y'),
+      ...require('./rules/a11y'),
     },
   },
 ]
