+++ lib/index.js
@@ -51,8 +51,8 @@ module.exports = [
       react: { version: reactVersion || '999.999.999' },
     },
     rules: {
-      ...require('./lib/rules/node'),
-      ...require('./lib/rules/react'),
+      ...require('./rules/node'),
+      ...require('./rules/react'),
     },
   },
   {
@@ -73,7 +73,7 @@ module.exports = [
     },
     // If adding a typescript-eslint version of an existing ESLint rule,
     // make sure to disable the ESLint rule here.
-    rules: require('./lib/rules/typescript'),
+    rules: require('./rules/typescript'),
   },
   {
     files: ['**/*.d.ts'],
