+++ lib/jest-recommended.js
