+++ lib/jest.js
@@ -9,7 +9,7 @@ module.exports = [
     languageOptions: jest.configs['flat/recommended'].languageOptions,
     rules: {
       ...jest.configs['flat/recommended'].rules,
-      ...require('./lib/rules/jest'),
+      ...require('./rules/jest'),
     },
   },
 ]
