+++ lib/node-recommended.js
