+++ lib/node.js
@@ -22,7 +22,7 @@ module.exports = [
         ...globals.node,
       },
     },
-    rules: require('./lib/rules/node'),
+    rules: require('./rules/node'),
   },
   {
     files: ['**/*.{ts,mts,cts}'],
@@ -39,6 +39,6 @@ module.exports = [
     },
     // If adding a typescript-eslint version of an existing ESLint rule,
     // make sure to disable the ESLint rule here.
-    rules: require('./lib/rules/typescript'),
+    rules: require('./rules/typescript'),
   },
 ]
