+++ lib/react-recommended.js
