+++ package.json
@@ -15,7 +15,26 @@
     "eslintconfig",
     "eslint-config"
   ],
-  "main": "index.js",
+  "main": "lib/index.js",
+  "files": [
+    "lib"
+  ],
+  "exports": {
+    ".": "./lib/index.js",
+    "./a11y": "./lib/a11y.js",
+    "./custom": "./lib/custom.js",
+    "./jest": "./lib/jest.js",
+    "./jest-recommended": "./lib/jest-recommended.js",
+    "./jsx-runtime": "./lib/jsx-runtime.js",
+    "./mdx": "./lib/mdx.js",
+    "./node": "./lib/node.js",
+    "./node-recommended": "./lib/node-recommended.js",
+    "./prettier": "./lib/prettier.js",
+    "./react-recommended": "./lib/react-recommended.js",
+    "./testing-library": "./lib/testing-library.js",
+    "./vitest": "./lib/vitest.js",
+    "./vitest-recommended": "./lib/vitest-recommended.js"
+  },
   "scripts": {
     "preinstall": "npx only-allow pnpm",
     "commit": "git-cz",
@@ -34,46 +53,39 @@
     "test:cov": "vitest run --coverage"
   },
   "dependencies": {
-    "confusing-browser-globals": "^1.0.11",
-    "eslint-plugin-import-x": "^4.17.1",
-    "eslint-plugin-unused-imports": "^4.4.1",
-    "globals": "^17.8.0",
-    "lodash.defaultsdeep": "^4.6.1"
+    "confusing-browser-globals": "1.0.11",
+    "eslint-plugin-import-x": "4.17.1",
+    "eslint-plugin-unused-imports": "4.4.1",
+    "globals": "17.8.0",
+    "lodash.defaultsdeep": "4.6.1"
   },
   "devDependencies": {
-    "@commitlint/cli": "^19.5.0",
-    "@commitlint/config-conventional": "^19.5.0",
-    "@commitlint/cz-commitlint": "^19.5.0",
-    "@typescript-eslint/eslint-plugin": "^8.65.0",
-    "@typescript-eslint/parser": "^8.65.0",
-    "@vitest/eslint-plugin": "^1.6.0",
-    "commitizen": "^4.3.1",
-    "commitlint": "^19.5.0",
-    "eslint": "^10.8.0",
-    "eslint-config-prettier": "^10.1.8",
-    "eslint-plugin-jest": "^29.16.0",
-    "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-mdx": "^3.8.1",
-    "eslint-plugin-prettier": "^5.5.0",
-    "eslint-plugin-react": "^7.37.5",
-    "eslint-plugin-react-hooks": "^7.1.1",
-    "eslint-plugin-testing-library": "^7.16.2",
-    "git-cz": "^4.9.0",
-    "husky": "^9.1.6",
-    "inquirer": "^9.3.7",
-    "jest": "^30.4.2",
-    "lint-staged": "^15.2.10",
-    "pinst": "^3.0.0",
-    "prettier": "^3.3.3",
-    "typescript": "^5.7.3",
-    "vitest": "^3.0.0"
-  },
-  "pnpm": {
-    "overrides": {
-      "normalize-package-data@2.5.0>semver": "^5.7.2",
-      "remark-parse@8.0.3>trim": "^0.0.3",
-      "@typescript-eslint/utils": "^8.65.0"
-    }
+    "@commitlint/cli": "21.2.1",
+    "@commitlint/config-conventional": "21.2.0",
+    "@commitlint/cz-commitlint": "21.2.0",
+    "@typescript-eslint/eslint-plugin": "8.65.0",
+    "@typescript-eslint/parser": "8.65.0",
+    "@vitest/eslint-plugin": "1.6.24",
+    "commitizen": "4.3.2",
+    "commitlint": "21.2.1",
+    "eslint": "10.8.0",
+    "eslint-config-prettier": "10.1.8",
+    "eslint-plugin-jest": "29.16.0",
+    "eslint-plugin-jsx-a11y": "6.10.2",
+    "eslint-plugin-mdx": "3.8.1",
+    "eslint-plugin-prettier": "5.5.6",
+    "eslint-plugin-react": "7.37.5",
+    "eslint-plugin-react-hooks": "7.1.1",
+    "eslint-plugin-testing-library": "7.16.2",
+    "git-cz": "4.9.0",
+    "husky": "9.1.7",
+    "inquirer": "14.0.2",
+    "jest": "30.4.2",
+    "lint-staged": "17.3.0",
+    "pinst": "3.0.0",
+    "prettier": "3.9.6",
+    "typescript": "6.0.3",
+    "vitest": "4.1.10"
   },
   "peerDependencies": {
     "@typescript-eslint/eslint-plugin": "^8.0.0",
@@ -89,7 +101,7 @@
     "eslint-plugin-react-hooks": "^7.0.0",
     "eslint-plugin-testing-library": "^7.0.0",
     "prettier": "^3.0.0",
-    "typescript": "^4 || ^5"
+    "typescript": "^5 || ^6"
   },
   "peerDependenciesMeta": {
     "@typescript-eslint/eslint-plugin": {
@@ -172,5 +184,5 @@
       "@semantic-release/github"
     ]
   },
-  "packageManager": "pnpm@9.1.4"
+  "packageManager": "pnpm@11.18.0"
 }
