+++ vitest.config.js
@@ -2,24 +2,20 @@ module.exports = {
   test: {
     globals: true,
     environment: 'node',
-    testTimeout: 10000,
     projects: [
       {
         test: {
           name: 'UNIT',
           globals: true,
-          environment: 'node',
-          include: ['**/*.spec.js', '**/*.test.js'],
-          exclude: ['**/node_modules/**', '**/*.rule.spec.js', '**/*.rule.test.js'],
+          include: ['**/*.spec.js'],
+          exclude: ['**/node_modules/**', '**/*.rule.spec.js'],
         },
       },
       {
         test: {
           name: 'ESLint',
           globals: true,
-          environment: 'node',
-          testTimeout: 10000,
-          include: ['**/*.rule.spec.js', '**/*.rule.test.js'],
+          include: ['**/*.rule.spec.js'],
           exclude: ['**/node_modules/**'],
         },
       },
